+++ inlang/source-code/paraglide/paraglide-next/src/app/middleware/detection/CookieDetection.ts
@@ -4,7 +4,7 @@ import { LANG_COOKIE } from "../../constants"
 export const createCookieDetection =
 	<T extends string>(cfg: { availableLanguageTags: readonly T[] }): LanguageDetector<T> =>
 	(request) => {
-		const localeCookeValue = request.cookies.get(LANG_COOKIE.name)?.value
-		if (cfg.availableLanguageTags.includes(localeCookeValue as T)) return localeCookeValue as T
+		const localeCookieValue = request.cookies.get(LANG_COOKIE.name)?.value
+		if (cfg.availableLanguageTags.includes(localeCookieValue as T)) return localeCookieValue as T
 		else return undefined
 	}