+++ inlang/source-code/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/cli
 
+## 2.16.5
+
+### Patch Changes
+
+- Updated dependencies [5209b81]
+  - @inlang/sdk@0.34.7
+
+## 2.16.4
+
+### Patch Changes
+
+- Updated dependencies [f38536e]
+  - @inlang/sdk@0.34.6
+
 ## 2.16.3
 
 ### Patch Changes