+++ inlang/source-code/sdk/multi-project-test/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Patch Changes
 
+- @inlang/cli@2.16.5
+
+## null
+
+### Patch Changes
+
+- @inlang/cli@2.16.4
+
+## null
+
+### Patch Changes
+
 - @inlang/cli@2.16.3
 
 ## null